Added csv output
This patch adds csv-output (/ge csv). First line is headline. Separator is ';'. ';' in text is converted to ','.
note: The xml-output (/ge export) is very memoryhungry, if the user is in a large guild, it may crash the game (stack overflow).
I removed whitespace from xml output in a previous patch to help with this issue.
Using csv-output should use a lot less ram, also output file is approx. 30-35% xml-file size.